+++ src/main/scala/edu/ie3/simona/model/participant/PrimaryDataParticipantModel.scala
@@ -19,7 +19,7 @@ import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.ontology.messages.flex.{
   FlexOptions,
   FlexType,
-  MinMaxFlexOptions,
+  PowerLimitFlexOptions,
 }
 import edu.ie3.simona.service.Data.PrimaryData.{
   ComplexPower,
@@ -63,7 +63,7 @@ final case class PrimaryDataParticipantModel[PD <: PrimaryData: ClassTag](
 
   override val flexModels
       : Map[FlexType, ParticipantFlexModel[PrimaryDataState[PD]]] = Map(
-    FlexType.MinMax -> PrimaryDataMinMaxFlexModel(this)
+    FlexType.PowerLimit -> PrimaryDataPowerLimitFlexModel(this)
   )
 
   override def determineState(
@@ -185,7 +185,7 @@ object PrimaryDataParticipantModel {
     override val reactivePower: Option[ReactivePower] = None
   }
 
-  private final case class PrimaryDataMinMaxFlexModel[PD <: PrimaryData](
+  private final case class PrimaryDataPowerLimitFlexModel[PD <: PrimaryData](
       model: PrimaryDataParticipantModel[PD]
   ) extends ParticipantFlexModel[PrimaryDataState[PD]] {
 
@@ -194,7 +194,7 @@ object PrimaryDataParticipantModel {
     ): FlexOptions = {
       val (operatingPoint, _) = model.determineOperatingPoint(state)
 
-      MinMaxFlexOptions.noFlexOption(operatingPoint.activePower)
+      PowerLimitFlexOptions.noFlexOption(operatingPoint.activePower)
     }
 
   }