+++ src/test/scala/edu/ie3/simona/model/em/EmAggregatePowerOptSpec.scala
@@ -7,7 +7,7 @@
 package edu.ie3.simona.model.em
 
 import edu.ie3.datamodel.models.input.system.{PvInput, SystemParticipantInput}
-import edu.ie3.simona.ontology.messages.flex.MinMaxFlexOptions
+import edu.ie3.simona.ontology.messages.flex.PowerLimitFlexOptions
 import edu.ie3.simona.test.common.UnitSpec
 import org.scalatestplus.mockito.MockitoSugar
 import squants.energy.Kilowatts
@@ -18,7 +18,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     val strat = EmAggregatePowerOpt(curtailRegenerative = true)
 
     "work with single flex options" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(4.0),
@@ -30,14 +30,14 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(0.0),
         Kilowatts(-1.0),
         Kilowatts(4.0),
       )
     }
     "work as expected at zero flexibility" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(5.0),
         min = Kilowatts(5.0),
         max = Kilowatts(5.0),
@@ -49,7 +49,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(5.0),
         Kilowatts(5.0),
         Kilowatts(5.0),
@@ -61,13 +61,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     val strat = EmAggregatePowerOpt(curtailRegenerative = true)
 
     "pick 0kW if possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(4.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -80,7 +80,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(0.0),
         Kilowatts(-7.0),
         Kilowatts(4.0),
@@ -88,13 +88,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick minSum if minSum > 0kW" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(6.0),
         min = Kilowatts(4.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-2.0),
         max = Kilowatts(2.0),
@@ -107,7 +107,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(2.0),
         Kilowatts(2.0),
         Kilowatts(14.0),
@@ -115,13 +115,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick maxSum if maxSum < 0kW" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(-1.0),
         min = Kilowatts(-10.0),
         max = Kilowatts(-1.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -134,7 +134,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-1.0),
         Kilowatts(-16.0),
         Kilowatts(-1.0),
@@ -146,13 +146,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     val strat = EmAggregatePowerOpt(curtailRegenerative = false)
 
     "exclude PV max power when normally picking 0kW as target" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(4.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -165,7 +165,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-2.0),
         Kilowatts(-7.0),
         Kilowatts(4.0),
@@ -173,13 +173,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "exclude PV max power when normally picking maxSum as target" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(-1.0),
         min = Kilowatts(-10.0),
         max = Kilowatts(-1.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -192,7 +192,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-7.0),
         Kilowatts(-16.0),
         Kilowatts(-1.0),
@@ -205,13 +205,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     val strat = EmAggregatePowerOpt(powerTarget, curtailRegenerative = true)
 
     "pick closed possible power if power target is not possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(5.0),
         min = Kilowatts(4.0),
         max = Kilowatts(6.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(0.0),
         min = Kilowatts(0.0),
         max = Kilowatts(0.0),
@@ -224,7 +224,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(4.0),
         Kilowatts(4.0),
         Kilowatts(6.0),
@@ -232,13 +232,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "use min flex of to stay inside inside power target limits" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(10.0),
         min = Kilowatts(9.0),
         max = Kilowatts(11.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -251,7 +251,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(3.0),
         Kilowatts(3.0),
         Kilowatts(11.0),
@@ -259,13 +259,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "use ref and stay inside inside power target limits" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(8.0),
         min = Kilowatts(7.0),
         max = Kilowatts(11.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -278,7 +278,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(2.0),
         Kilowatts(1.0),
         Kilowatts(11.0),
@@ -286,13 +286,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick power target if possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(9.0),
         min = Kilowatts(8.0),
         max = Kilowatts(10.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -305,7 +305,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(3.0),
         Kilowatts(2.0),
         Kilowatts(10.0),
@@ -313,13 +313,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick reference inside power target limits if possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(9.0),
         min = Kilowatts(8.0),
         max = Kilowatts(10.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-7.0),
         min = Kilowatts(-7.0),
         max = Kilowatts(0.0),
@@ -332,7 +332,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(2.0),
         Kilowatts(1.0),
         Kilowatts(10.0),
@@ -340,13 +340,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "stay inside inside power target limits and not reduce renewable generation" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(9.0),
         min = Kilowatts(8.0),
         max = Kilowatts(10.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-12.0),
         min = Kilowatts(-12.0),
         max = Kilowatts(0.0),
@@ -359,7 +359,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-3.0),
         Kilowatts(-4.0),
         Kilowatts(10.0),
@@ -367,13 +367,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "use max flex of to stay inside inside power target limits and not reduce renewable generation" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(1.0),
         min = Kilowatts(0.0),
         max = Kilowatts(5.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-8.0),
         min = Kilowatts(-8.0),
         max = Kilowatts(0.0),
@@ -386,7 +386,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-3.0),
         Kilowatts(-8.0),
         Kilowatts(5.0),
@@ -394,13 +394,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick 3kW if possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(10.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -413,7 +413,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(3.0),
         Kilowatts(-7.0),
         Kilowatts(12.0),
@@ -421,13 +421,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick -3kW if possible" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(4.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -440,7 +440,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-3.0),
         Kilowatts(-7.0),
         Kilowatts(4.0),
@@ -448,13 +448,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick reference power (positive) if inside the power target limits" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(8.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(10.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -467,7 +467,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(2.0),
         Kilowatts(-7.0),
         Kilowatts(10.0),
@@ -475,13 +475,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick (negative) reference power if inside the power target limits" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(4.0),
         min = Kilowatts(-1.0),
         max = Kilowatts(4.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -494,7 +494,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-2.0),
         Kilowatts(-7.0),
         Kilowatts(4.0),
@@ -502,13 +502,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick minSum if minSum > 0kW" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(6.0),
         min = Kilowatts(4.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-2.0),
         max = Kilowatts(2.0),
@@ -521,7 +521,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(3.0),
         Kilowatts(2.0),
         Kilowatts(14.0),
@@ -529,13 +529,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick maxSum if maxSum < 0kW" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(-1.0),
         min = Kilowatts(-10.0),
         max = Kilowatts(-1.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -548,7 +548,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-3.0),
         Kilowatts(-16.0),
         Kilowatts(-1.0),
@@ -561,13 +561,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     val strat = EmAggregatePowerOpt(powerTarget, curtailRegenerative = false)
 
     "pick min power to get closed possible to power target if it cannot reached" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(11.0),
         min = Kilowatts(10.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -580,7 +580,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(4.0),
         Kilowatts(4.0),
         Kilowatts(12.0),
@@ -588,13 +588,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick reference power if power target can be reached" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(9.0),
         min = Kilowatts(8.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -607,7 +607,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(3.0),
         Kilowatts(2.0),
         Kilowatts(12.0),
@@ -615,13 +615,13 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
     }
 
     "pick max power to get closed possible to power target if it cannot reached" in {
-      val flexOptions1 = MinMaxFlexOptions(
+      val flexOptions1 = PowerLimitFlexOptions(
         ref = Kilowatts(1.0),
         min = Kilowatts(0.0),
         max = Kilowatts(2.0),
       )
 
-      val flexOptions2 = MinMaxFlexOptions(
+      val flexOptions2 = PowerLimitFlexOptions(
         ref = Kilowatts(-6.0),
         min = Kilowatts(-6.0),
         max = Kilowatts(0.0),
@@ -634,7 +634,7 @@ class EmAggregatePowerOptSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe MinMaxFlexOptions(
+      actualResult shouldBe PowerLimitFlexOptions(
         Kilowatts(-4.0),
         Kilowatts(-6.0),
         Kilowatts(2.0),